Always set panel animation duration
The condition applied here will not apply the animation duration when
it's set at 0. However, this means that the duration will be set to the
default duration (1.0), making the panel disrespect the animation when
it's disabled. Fix this by removing the condition entirely and instead,
fix the jumpy view issue by modifying settleCapturedViewAt().